+++ public/js/index.js
@@ -1,70 +1,88 @@
+const keyword = ""
+const page = 0
+const attractionUrl = "/api/attractions";
+
+//驗證是否登入
+fetch("/api/user/auth",{
+    method: "GET"
+})
+.then(function(response){
+    return response.json();
+})
+.then(function(data){
+    let signUp = document.getElementById("func");
+    if (data.data != null){
+        signUp.setAttribute("onclick","signup()")
+        signUp.textContent = "登出系統"
+    }else{
+        signUp.textContent = "登入/註冊"
+    };
+});
+
 //首頁載入景點
-let keyword = ""
-let page = 0
-let url = "/api/attractions";
-    fetch(url+"?page="+page)
-    .then(function(response){
-        return response.json();
-    })
-    .then(function(data){
-        let dataList = data.data;
-        let firstImg = [];
-        let name = [];
-        let mrtContent = [];
-        let categoryContent = [];
-        let id = [];
-        for (let i = 0; i<data.data.length; i++){
-            firstImg.push(dataList[i].images[0]);
-            name.push(dataList[i].name);
-            mrtContent.push(dataList[i].mrt);
-            categoryContent.push(dataList[i].category);
-            id.push(dataList[i].id);
-        };
-        let target = document.querySelectorAll(".attraction_img");
-        let transportTarget = document.querySelectorAll(".transport");                
-        let oldImg = document.querySelectorAll(".img");
-        let attraction_name = document.querySelectorAll(".attraction_name");
-        let mrt = document.querySelectorAll(".mrt");
-        let category = document.querySelectorAll(".category");
-        for (let i = 0; i<12; i++){
-           //img
-            let newImg = document.createElement("img");
-            newImg.className = "img";
-            newImg.src = firstImg[i];
-            target[i].replaceChild(newImg,oldImg[i]);
-            //name
-            let newName = document.createElement("div");
-            newName.className = "attraction_name";
-            let nameTextNode = document.createTextNode(name[i]);
-            newName.appendChild(nameTextNode);
-            target[i].replaceChild(newName,attraction_name[i]);
-            //mrt
-            let newMrt = document.createElement("div");
-            newMrt.className = "mrt";
-            let mrtTextNode = document.createTextNode(mrtContent[i]);
-            newMrt.appendChild(mrtTextNode);
-            transportTarget[i].replaceChild(newMrt,mrt[i]); 
-            //category
-            let newCategory = document.createElement("div");
-            newCategory.className = "category";
-            let categoryTextNode = document.createTextNode(categoryContent[i]);
-            newCategory.appendChild(categoryTextNode);
-            transportTarget[i].replaceChild(newCategory,category[i]);  
-            //跳轉景點頁面
-            let link = document.querySelectorAll(".link");
-            for(let i = 0; i<link.length; i++){
-                link[i].setAttribute("href",`/attraction/${id[i]}`);
-            }
-        };
-    });
+fetch(attractionUrl+"?page="+page)
+.then(function(response){
+    return response.json();
+})
+.then(function(data){
+    let dataList = data.data;
+    let firstImg = [];
+    let name = [];
+    let mrtContent = [];
+    let categoryContent = [];
+    let id = [];
+    for (let i = 0; i<data.data.length; i++){
+        firstImg.push(dataList[i].images[0]);
+        name.push(dataList[i].name);
+        mrtContent.push(dataList[i].mrt);
+        categoryContent.push(dataList[i].category);
+        id.push(dataList[i].id);
+    };
+    let target = document.querySelectorAll(".attraction_img");
+    let transportTarget = document.querySelectorAll(".transport");                
+    let oldImg = document.querySelectorAll(".img");
+    let attraction_name = document.querySelectorAll(".attraction_name");
+    let mrt = document.querySelectorAll(".mrt");
+    let category = document.querySelectorAll(".category");
+    for (let i = 0; i<12; i++){
+        //img
+        let newImg = document.createElement("img");
+        newImg.className = "img";
+        newImg.src = firstImg[i];
+        target[i].replaceChild(newImg,oldImg[i]);
+        //name
+        let newName = document.createElement("div");
+        newName.className = "attraction_name";
+        let nameTextNode = document.createTextNode(name[i]);
+        newName.appendChild(nameTextNode);
+        target[i].replaceChild(newName,attraction_name[i]);
+        //mrt
+        let newMrt = document.createElement("div");
+        newMrt.className = "mrt";
+        let mrtTextNode = document.createTextNode(mrtContent[i]);
+        newMrt.appendChild(mrtTextNode);
+        transportTarget[i].replaceChild(newMrt,mrt[i]); 
+        //category
+        let newCategory = document.createElement("div");
+        newCategory.className = "category";
+        let categoryTextNode = document.createTextNode(categoryContent[i]);
+        newCategory.appendChild(categoryTextNode);
+        transportTarget[i].replaceChild(newCategory,category[i]);  
+        //跳轉景點頁面
+        let link = document.querySelectorAll(".link");
+        for(let i = 0; i<link.length; i++){
+            link[i].setAttribute("href",`/attraction/${id[i]}`);
+        }
+    };
+});
 
 //無限滾輪
+let nextPage = 1
 let options = {
     root:null,
     rootMargin:"0px",
     threshold:0.1,
 };
-let nextPage = 1
 let callback = (entries,observer) => {
     entries.forEach(entry => {
         if (!entry.isIntersecting) {
@@ -73,7 +91,7 @@ let callback = (entries,observer) => {
         if (nextPage == null){
             return;
         }
-        fetch(url+"?page="+nextPage+"&keyword="+search.value)
+        fetch(attractionUrl+"?page="+nextPage+"&keyword="+search.value)
         .then(function(response){
             return response.json();
         })
@@ -121,7 +139,7 @@ observer.observe(target);
 //關鍵字搜尋
 let search = document.querySelector(".input_text");
 function searchData(){
-    fetch(url+"?page=0&keyword="+search.value)
+    fetch(attractionUrl+"?page=0&keyword="+search.value)
     .then(function(response){
         return response.json();
     })
@@ -171,7 +189,6 @@ function searchData(){
 };
 
 //關鍵字列表
-let input_text = document.querySelector(".input_text");
 let categories_menu = document.querySelector(".categories_menu");
 fetch("/api/categories")
 .then(function(response){
@@ -186,13 +203,13 @@ fetch("/api/categories")
         categories_list.appendChild(category_list);
     };
 });
-input_text.addEventListener("click",function(event){
+search.addEventListener("click",function(event){
     categories_menu.style.display = "block";
     event.stopPropagation();
     let category_list=document.querySelectorAll(".category_list");
     for(let i = 0; i<category_list.length; i++){
         category_list[i].addEventListener("click",function(){
-        input_text.value = category_list[i].textContent;
+        search.value = category_list[i].textContent;
         categories_menu.style.display = "none";    
         });
     };       
@@ -202,10 +219,11 @@ document.addEventListener("click",function(){
 });
 
 //註冊＆登入
-let signInContent = document.querySelector(".sign-in");
-let registerContent = document.querySelector(".register");
-let hideBg=document.querySelector(".hide-bg");
-let body = document.querySelector(".body");
+const signInContent = document.querySelector(".sign-in");
+const registerContent = document.querySelector(".register");
+const hideBg=document.querySelector(".hide-bg");
+const body = document.querySelector(".body");
+
 //打開登入頁面
 function sign(){
     signInContent.style.display = "block";
@@ -234,19 +252,19 @@ function toSignIn(){
 
 //登入會員
 function signIn(){
-    email = document.getElementById("signInEmail").value;
-    password = document.getElementById("signInPassword").value;
+    let signInEmail = document.getElementById("signInEmail").value;
+    let signInPassword = document.getElementById("signInPassword").value;
     data = {
-        "email":email,
-        "password":password
+        "email" : signInEmail,
+        "password" : signInPassword
     };
     fetch("/api/user/auth",{
         method: "PUT",
         body: JSON.stringify(data),
         cache: "no-cache",
         headers:{
-            "Accept":"application/json",
-            "Content-Type":"application/json"
+            "Accept" : "application/json",
+            "Content-Type" : "application/json"
         }  
     })
     .then(function(response){
@@ -254,47 +272,63 @@ function signIn(){
     })
     .then(function(res){
         if (res.ok == true) {
-            alert("登入成功");
-            signInContent.style.display = "none";
-            hideBg.style.display="none";
-            body.style.overflow = "auto"
+            window.location.reload();
         };
         if (res.error == true) {
-            alert(res.message);
+            let signAlert = document.querySelector(".sign-alert")
+            let sign = document.querySelector(".sign-in")
+            sign.style.height = "298px"
+            signAlert.textContent = res.message;
         };          
     });
 };
 
 //註冊會員
 function register(){
-    userName= document.getElementById("registerName").value;
-    email = document.getElementById("registerEmail").value;
-    password = document.getElementById("registerPassword").value;
+    let registerName = document.getElementById("registerName").value;
+    let registerEmail = document.getElementById("registerEmail").value;
+    let registerPassword = document.getElementById("registerPassword").value;
     data = {
-        "name":userName,
-        "email":email,
-        "password":password
+        "name" : registerName,
+        "email" : registerEmail,
+        "password" : registerPassword
     };
     fetch("/api/user",{
         method: "POST",
         body: JSON.stringify(data),
         cache: "no-cache",
         headers:{
-            "Accept":"application/json",
-            "Content-Type":"application/json"
+            "Accept" : "application/json",
+            "Content-Type" : "application/json"
         }  
     })
     .then(function(response){
         return response.json();
     })
     .then(function(res){
+        let registerAlert = document.querySelector(".register-alert")
+        let register = document.querySelector(".register")
+        register.style.height = "360px"
         if (res.ok == true) {
-            alert("註冊成功");
-            signInContent.style.display = "block";
-            registerContent.style.display = "none";
+            registerAlert.textContent = "註冊成功，請登入頁面"
         };
         if (res.error == true) {
-            alert(res.message);
+            registerAlert.textContent = res.message;
         };          
+    });
+};
+
+//登出系統
+function signup(){
+    fetch("/api/user/auth",{
+        method: "DELETE"
+    })
+    .then(function(response){
+        return response.json();
+    })
+    .then(function(res){
+        if (res.ok == true) {
+            window.location.reload();
+        };
     });
 };
